+++ src/Semaphore/SemaphoreTable.css
@@ -1,0 +1,8 @@
+.SemaphoreTable .table-responsive {
+  border-top: none;
+}
+  
+.SemaphoreTable th,
+.SemaphoreTable td {
+  text-align: center;
+}
