+++ src/Views/Encoding/Semaphore.tsx
@@ -1,16 +1,36 @@
 import * as React from 'react';
+import { Tab, Tabs } from 'react-bootstrap';
+import TabViewBase from '../TabViewBase';
 import SemaphoreStream from '../../Semaphore/SemaphoreStream';
+import SemaphoreTable from '../../Semaphore/SemaphoreTable';
 import './Semaphore.css';
 
-class Semaphore extends React.Component {
+class Semaphore extends TabViewBase {
   public render() {
     return (
       <div className="Semaphore">
         <div className="Semaphore-content">
-          <SemaphoreStream />
+          <Tabs
+            activeKey={this.state.activeKey}
+            animation={false}
+            id="Semaphore-tabs"
+            // tslint:disable-next-line: no-any
+            onSelect={(activeKey: any) => this.onTabSelect(activeKey as number)}
+          >
+            <Tab eventKey={1} title="Reference">
+              <SemaphoreTable />
+            </Tab>
+            <Tab eventKey={2} title="Value">
+              <SemaphoreStream />
+            </Tab>
+          </Tabs>
         </div>
       </div>
     );
+  }
+
+  protected getLocalStorageKey() {
+    return 'Semaphore';
   }
 }
