+++ src/Views/Reference/Resistors.tsx
@@ -1,29 +1,10 @@
 import * as React from 'react';
 import { Tab, Tabs } from 'react-bootstrap';
+import TabViewBase from '../TabViewBase';
 import ResistorTable from '../../Resistor/ResistorTable';
 import ResistorInput from '../../Resistor/ResistorInput';
-import LocalStorageComponent from '../../Data/LocalStorageComponent';
 
-type Props = {};
-type State = {
-  activeKey: number,
-};
-
-type SavedState = {
-  activeKey: number,
-};
-
-class Resistors extends LocalStorageComponent<Props, State, SavedState> {
-  private _activeKey: number = 1;
-
-  constructor(props: Props) {
-    super(props);
-
-    this.state = {
-      activeKey: this._activeKey,
-    };
-  }
-
+class Resistors extends TabViewBase {
   public render() {
     return (
       <div className="Resistors">
@@ -49,29 +30,6 @@ class Resistors extends LocalStorageComponent<Props, State, SavedState> {
 
   protected getLocalStorageKey() {
     return 'Resistors';
-  }
-
-  protected onSaveState() {
-    return {
-      activeKey: this._activeKey,
-    };
-  }
-
-  protected onRestoreState(savedState: SavedState | null) {
-    if (savedState !== null) {
-      this._activeKey = savedState.activeKey;
-    }
-  }
-
-  protected onUpdateState() {
-    this.setState({
-      activeKey: this._activeKey,
-    });
-  }
-
-  private onTabSelect(activeKey: number) {
-    this._activeKey = activeKey;
-    this.updateState();
   }
 }
